# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TogetherPlan API",
        "description": "API documentation for the TogetherPlan project. This API allows users to manage events, participants, voting on dates, groups, and notifications.",
        "contact": {
            "name": "TogetherPlan Support",
            "email": "support@togetherplan.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.togetherplan.com",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/groups": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Liste tous les groupes",
                "description": "Afficher tous les groupes",
                "operationId": "f286cbf34dc4be32a94459a335658168",
                "responses": {
                    "200": {
                        "description": "Liste des groupes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Group"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Créer un nouveau groupe",
                "description": "Créer un nouveau groupe",
                "operationId": "74e9878763e20c38ae5c9930b8d2d3f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Groupe Dev"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du groupe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Groupe créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/groups/{id}": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Afficher un groupe spécifique",
                "description": "Afficher un groupe spécifique",
                "operationId": "38a1f4990d0dc69ead7f7377ad016117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du groupe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Groupe non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Supprimer un groupe",
                "description": "Supprimer un groupe",
                "operationId": "8c693e934c350f47618655bc71886930",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groupe supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Groupe supprimé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/groups/{id}/invite": {
            "post": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Inviter des utilisateurs par email (ajout automatique au groupe)",
                "description": "Inviter des utilisateurs par email (ajout automatique au groupe)",
                "operationId": "fd86a54e9a7275a6bba8511b2b1b766a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du groupe",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emails"
                                ],
                                "properties": {
                                    "emails": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateurs ajoutés et invitations envoyées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateurs ajoutés et invitations envoyées !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Group": {
                "title": "Group",
                "required": [
                    "id",
                    "name",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Groupe Dev"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du groupe"
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Laravel Sanctum API token for authentication",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}